+++ test/app/lib/modules/test/usercommands/test.js
@@ -1,0 +1,8 @@
+exports.access = 'user';
+
+exports.params = [];
+
+exports.execute = function (state, callback) {
+	state.respond('test');
+	return callback();
+};